rename campaigns to marketing_campaigns
