+++ lib/sendgrid4r/factory/campaign_factory.rb
@@ -10,7 +10,7 @@ module SendGrid4r
         segment_ids: nil, categories: nil, suppression_group_id: nil,
         custom_unsubscribe_url: nil, ip_pool: nil, html_content: nil,
         plain_content: nil)
-        segment = SendGrid4r::REST::Campaigns::Campaigns::Campaign.new(
+        segment = SendGrid4r::REST::MarketingCampaigns::Campaign.new(
           nil, title, subject, sender_id, list_ids, segment_ids, categories,
           suppression_group_id, custom_unsubscribe_url, ip_pool, html_content,
           plain_content, nil, nil
